+++ src/3rdparty/super4pcs/src/super4pcs/algorithms/super4pcs.h
@@ -94,7 +94,7 @@ public:
          int base_point1,
          int base_point2,
          PairsVector* pairs, 
-         std::vector<float> ppf_) const override;
+         std::vector<int> ppf_) const override;
 
  // Finds congruent candidates in the set Q, given the invariants and threshold
  // distances. Returns true if a non empty set can be found, false otherwise.
