+++ src/physim_pose_estimation/src/hypothesis_generation/ObjectPoseCandidateSet.hpp
@@ -12,8 +12,8 @@ namespace pose_candidates{
 
 		void generate(std::string objName, std::string scenePath, 
 				pcl::PointCloud<pcl::PointXYZRGBNormal>::Ptr pclSegment, pcl::PointCloud<pcl::PointXYZRGBNormal>::Ptr pclModel, 
-				pcl::PointCloud<pcl::PointXYZRGBNormal>::Ptr pclModelSampled, std::map<std::vector<float>, float> PPFMap, 
-				float max_count_ppf, Eigen::Matrix3f camIntrinsic);
+				pcl::PointCloud<pcl::PointXYZRGBNormal>::Ptr pclModelSampled, std::map<std::vector<int>, int> PPFMap, 
+				int max_count_ppf, Eigen::Matrix3f camIntrinsic);
 		void cluster();
 
 		std::vector< std::pair <Eigen::Isometry3d, float> > hypothesisSet;
